Reorganize QtFRED file structure
