+++ qtfred/src/mission/FredRenderer.h
