+++ qtfred/src/mission/editor.h
