+++ qtfred/src/mission/iterators.h
