+++ qtfred/src/mission/object.cpp
