+++ qtfred/src/mission/object.h
