+++ qtfred/src/ui/QtGraphicsOperations.cpp
