+++ qtfred/src/ui/fredGlobals.cpp
