+++ qtfred/src/ui/fredGlobals.h
