+++ qtfred/src/ui/mainwindow.h
