+++ qtfred/src/ui/renderwidget.h
